Done all the changes
